fix: charge int type

+++ src/mlipaudit/benchmark.py
@@ -31,7 +31,7 @@ RunModeAsString: TypeAlias = Literal["dev", "fast", "standard"]
 
 #: Default total charge applied to `atoms.info["charge"]` when a benchmark's
 #: input data does not specify one. Represents a neutral system.
-DEFAULT_CHARGE: float = 0.0
+DEFAULT_CHARGE: int = 0
 
 #: Default spin multiplicity applied to `atoms.info["spin"]`. All benchmarks
 #: currently treat systems as closed-shell singlets unless overridden.
